Use a class to store configuration variables
Created a class that will be used to store configuration
variables. Changed blog's __init__ module to use this class.

Added a SECRET_KEY so that Flask is able to generate signatures
and tokens to protect against CSRF attacks. Can be gotten from
environment variable or hard coded inside the config module.