Add SQLALCHEMY_DATABASE_URI, SQLALCHEMY_TRACK_MODIFICATIONS
Added SQLALCHEMY_DATABASE_URI variable that stores the value of
the database location. It defaults to using SQLite if there's
no DATABASE_URL environment variable set.

Added SQLALCHEMY_TRACK_MODIFICATIONS and set it to False since
for now we do not want signals when a change is about to be made
to the database.