ft: Add register, login, logout views; require users to login
Added a view for registering new users.

Modify the login view to use real users that are stored in
a database. Only show the login view to users who aren't logged
in; return logged in users to the index.

Required users to login to view the index page and added logic
that redirect users upon login to the page they were trying to
access.

Removed the dummy user that was being passed to the template
since the current user can be accessed via Flask-Login.