ft: Add custom error handlers
Added custom error handlers with templates and error codes
Error handler for the 500 errors could be invoked after a database error,
so to make sure any failed database sessions do not interfere with any database
accesses triggered by the template, issue a session rollback.
This resets the session to a clean state.